+++ packages/terra-slide-panel/tests/wdio/slide-panel-spec.js
@@ -113,13 +113,15 @@ Terra.describeViewports('Slide panel', ['large'], () => {
       browser.pause(150);
       expect(browser.getText('#test-slide [aria-hidden="false"]')).to.equal('Increase Count 1');
 
+      Terra.validates.screenshot('panel with end position', { selector: '#root' });
+
       browser.click('#test-toggle');
       browser.waitForExist('#test-slide [aria-hidden="false"]');
       browser.pause(150);
 
       expect(browser.getText('#test-slide [aria-hidden="false"]')).to.equal('Increase Count 1');
 
-      Terra.validates.screenshot({ selector: '#root' });
+      Terra.validates.screenshot('panel with start position', { selector: '#root' });
     });
   });
 });