+++ packages/terra-slide-panel/CHANGELOG.md
@@ -3,8 +3,8 @@ Changelog
 
 Unreleased
 ----------
-### Changed
-* Gave main content and panel content keys to prevent unmounting when changing panel position
+### Fixed
+* Added key to main content and panel content to prevent unmounting when changing panel position
 
 3.23.0 - (February 18, 2020)
 ------------------